Refactor Jinja extensions into regular functions
